+++ apps/moq_lite.zig
@@ -696,7 +696,7 @@ pub fn main(init: std.process.Init.Minimal) !u8 {
     var cert_path: []const u8 = "interop/certs/server.crt";
     var key_path: []const u8 = "interop/certs/server.key";
 
-    var args = std.process.Args.Iterator.init(init.args);
+    var args = sys.argsIterator(init.args);
     _ = args.next();
     while (args.next()) |arg| {
         if (std.mem.eql(u8, arg, "publish")) {