+++ apps/quic_client.zig
@@ -45,7 +45,7 @@ pub fn main(init: std.process.Init.Minimal) !void {
 
     var port: u16 = 4434;
     var insecure = false;
-    var args = std.process.Args.Iterator.init(init.args);
+    var args = quic.sys.argsIterator(init.args);
     _ = args.next();
     while (args.next()) |arg| {
         if (std.mem.eql(u8, arg, "--port")) {